[HIB-128] Integrate ssm previous time with find_new lambda
