+++ lambdas/find_new_granules/src/find_new/find_new_env.py
@@ -4,6 +4,7 @@ class Environment:
         self.maturity = 'test'
         self.bucket = 'hyp3-in-a-box'
         self.scheduler_lambda = 'hyp3-scheduler-lambda'
+        self.ssm_previous_time_name = None
 
 
 environment = Environment()
