+++ lambdas/find_new_granules/tests/find_new_mocks.py
@@ -2,7 +2,6 @@ import json
 import pathlib as pl
 
 import import_find_new
-from find_new.previous_time import get_time_file_path
 
 
 def asf_api_requests_get(*args, **kwargs):
@@ -22,13 +21,3 @@ def is_api_url(url):
         "https://api.daac.asf.alaska.edu/services/search/param",
         "https://cmr.earthdata.nasa.gov/search/granules.json"
     )
-
-
-def get_s3_download_func(time):
-    def s3_download(*args, **kwargs):
-        time_dict = json.dumps({'previous': time.timestamp()})
-
-        with open(get_time_file_path(), 'w') as f:
-            f.write(time_dict)
-
-    return s3_download
